Move types to central location
